feat: do not normalize the billing anchor
As normalizing the billing anchor could move the anchor a few days (if
it is on > 28th day of a month)
